+++ tests/APITest.php
@@ -22,24 +22,9 @@
  * @link https://github.com/bheisig/i-doit-api-client-php
  */
 
-use PHPUnit\Framework\TestCase;
 use bheisig\idoitapi\API;
 
-class APITest extends TestCase {
-
-    /**
-     * @var \bheisig\idoitapi\API
-     */
-    protected $api;
-
-    public function setUp() {
-        $this->api = new API([
-            'url' => $GLOBALS['url'],
-            'key' => $GLOBALS['key'],
-            'username' => $GLOBALS['username'],
-            'password' => $GLOBALS['password']
-        ]);
-    }
+class APITest extends BaseTest {
 
     public function testConnect() {
         $this->assertInstanceOf(API::class, $this->api->connect());
@@ -109,13 +94,15 @@ class APITest extends TestCase {
     }
 
     public function testBatchRequest() {
+        $objectID = $this->createObject();
+
         $results = $this->api->batchRequest([
             [
                 'method' => 'idoit.version'
             ],
             [
                 'method' => 'cmdb.object.read',
-                'params' => ['id' => 1]
+                'params' => ['id' => $objectID]
             ]
         ]);