+++ tests/CMDBCategoryInfoTest.php
@@ -22,16 +22,9 @@
  * @link https://github.com/bheisig/i-doit-api-client-php
  */
 
-use PHPUnit\Framework\TestCase;
-use bheisig\idoitapi\API;
 use bheisig\idoitapi\CMDBCategoryInfo;
 
-class CMDBCategoryInfoTest extends TestCase {
-
-    /**
-     * @var \bheisig\idoitapi\API
-     */
-    protected $api;
+class CMDBCategoryInfoTest extends BaseTest {
 
     /**
      * @var \bheisig\idoitapi\CMDBCategoryInfo
@@ -41,12 +34,7 @@ class CMDBCategoryInfoTest extends TestCase {
     protected $categories = [];
 
     public function setUp() {
-        $this->api = new API([
-            'url' => $GLOBALS['url'],
-            'key' => $GLOBALS['key'],
-            'username' => $GLOBALS['username'],
-            'password' => $GLOBALS['password']
-        ]);
+        parent::setUp();
 
         $this->instance = new CMDBCategoryInfo($this->api);