+++ tests/ImageTest.php
@@ -22,16 +22,9 @@
  * @link https://github.com/bheisig/i-doit-api-client-php
  */
 
-use PHPUnit\Framework\TestCase;
-use bheisig\idoitapi\API;
 use bheisig\idoitapi\Image;
 
-class ImageTest extends TestCase {
-
-    /**
-     * @var \bheisig\idoitapi\API
-     */
-    protected $api;
+class ImageTest extends BaseTest {
 
     /**
      * @var \bheisig\idoitapi\Image
@@ -44,12 +37,7 @@ class ImageTest extends TestCase {
     protected $files = [];
 
     public function setUp() {
-        $this->api = new API([
-            'url' => $GLOBALS['url'],
-            'key' => $GLOBALS['key'],
-            'username' => $GLOBALS['username'],
-            'password' => $GLOBALS['password']
-        ]);
+        parent::setUp();
 
         $this->instance = new Image($this->api);
 
@@ -60,18 +48,22 @@ class ImageTest extends TestCase {
     }
 
     public function testAdd() {
+        $objectID = $this->createObject();
+
         foreach ($this->files as $filePath => $caption) {
             $this->assertInstanceOf(
                 Image::class,
-                $this->instance->add(9, $filePath, $caption)
+                $this->instance->add($objectID, $filePath, $caption)
             );
         }
     }
 
     public function testBatchAdd() {
+        $objectID = $this->createObject();
+
         $this->assertInstanceOf(
             Image::class,
-            $this->instance->batchAdd(10, $this->files)
+            $this->instance->batchAdd($objectID, $this->files)
         );
     }